Split submission button part of form into separate section to creator inputs
